Fixed resolve meta data method when saving received song from radio to temp mp3 file

+++ SmartRadio/SmartRadio.Data/Models/SongFingerprint.cs
@@ -17,6 +17,6 @@ namespace SmartRadio.Data.Models
 
         public long Hash { get; set; }
 
-        public short Offset { get; set; }
+        public int Offset { get; set; }
     }
 }
